refactor: API endpoint와 path 상수 파일 분리
