+++ src/constants/api.ts
@@ -1,0 +1,3 @@
+export const BASE_URL: string = "https://api.themoviedb.org/3";
+
+export const API_KEY: string = import.meta.env.VITE_TMDB_API_KEY;
