--- src/constants/constant.ts
@@ -1,12 +1,0 @@
-export const BACKDROP_IMAGE_URL: string =
-  "https://image.tmdb.org/t/p/w1920_and_h800_multi_faces/";
-
-export const THUMB_NAIL_URL: string = "https://media.themoviedb.org/t/p/w200";
-
-export const API_KEY: string = import.meta.env.VITE_TMDB_API_KEY;
-
-export const BASE_URL: string = "https://api.themoviedb.org/3";
-
-export const POPULAR_PATH: string = "movie/popular";
-
-export const SEARCH_PATH: string = "search/movie";
