+++ src/constants/image.ts
@@ -1,0 +1,4 @@
+export const BACKDROP_IMAGE_URL: string =
+  "https://image.tmdb.org/t/p/w1920_and_h800_multi_faces/";
+
+export const THUMB_NAIL_URL: string = "https://media.themoviedb.org/t/p/w200";
