+++ src/constants/path.ts
@@ -1,0 +1,3 @@
+export const POPULAR_PATH: string = "movie/popular";
+
+export const SEARCH_PATH: string = "search/movie";
